update buddy implementation

+++ network/socket.cpp
@@ -51,8 +51,8 @@ bool Socket::connect(const char *host, const char *service, int family, int sock
     return false;
 }
 
-int Socket::recv(void *buf, size_t size, int flags) {
-    int n;
+size_t Socket::recv(void *buf, size_t size, int flags) {
+    size_t n;
     while (true) {
 	    n = ::recv(_socket, buf, size, flags);
         if (!(n < 0 && errno == EINTR)) {
@@ -63,8 +63,8 @@ int Socket::recv(void *buf, size_t size, int flags) {
 	return n;
 }
 
-int Socket::send(const void *buf, size_t size, int flags) {
-    int n;
+size_t Socket::send(const void *buf, size_t size, int flags) {
+    size_t n;
     while (true) {
 	    n = ::send(_socket, buf, size, flags);
         if (!(n < 0 && errno == EINTR)) {
@@ -75,13 +75,13 @@ int Socket::send(const void *buf, size_t size, int flags) {
 	return n;
 }
 
-int Socket::send(const void *buf1, size_t size1, const void *buf2, size_t size2) {
+size_t Socket::send(const void *buf1, size_t size1, const void *buf2, size_t size2) {
     struct iovec iov[2];
     iov[0].iov_base = (char*)buf1;
     iov[0].iov_len = size1;
     iov[1].iov_base = (char*)buf2;
     iov[1].iov_len = size2;
-    int n;
+    size_t n;
 
     while (true) {
         n = writev(_socket, iov, 2);
@@ -93,7 +93,7 @@ int Socket::send(const void *buf1, size_t size1, const void *buf2, size_t size2)
     return n;
 }
 
-int Socket::send(const void *buf1, size_t size1, const void *buf2, size_t size2, const void *buf3, size_t size3) {
+size_t Socket::send(const void *buf1, size_t size1, const void *buf2, size_t size2, const void *buf3, size_t size3) {
     struct iovec iov[3];
     iov[0].iov_base = (char*)buf1;
     iov[0].iov_len = size1;
@@ -101,7 +101,7 @@ int Socket::send(const void *buf1, size_t size1, const void *buf2, size_t size2,
     iov[1].iov_len = size2;
     iov[2].iov_base = (char*)buf3;
     iov[2].iov_len = size3;
-    int n;
+    size_t n;
 
     while (true) {
         n = writev(_socket, iov, 3);
